Project settings

+++ balchug-engine/src/settings.rs
@@ -1,12 +1,12 @@
 #[derive(Copy, Clone)]
 pub struct Settings {
-    pub background_color: [f32;3],
+    pub background_color: [u8;3],
 }
 
 impl Default for Settings {
     fn default() -> Self {
         Self {
-            background_color: [0.0, 0.0, 0.0],
+            background_color: [0, 0, 0],
         }
     }
 }
